This PR implements a new backend supporting PostgreSQL and CockroachDB SQL database platforms.

Implements #10253 without cloud connectivity.