@jimbishopp 2022-03-16
One final UX feature I'd like to add: attempt to create the database if it does not exist. Should be easy: trap connection error saying DB doesn't exist, connect to "postgres" database, create "teleport" database, restart and connect to "teleport".