@jimbishopp 2022-03-22
Latest changes include an optimization for the DeleteItems query. 

Previous query:
https://github.com/gravitational/teleport/blob/1abf0b6b080d05a1110be4238cda79137e6e5b23/lib/backend/postgres/tx.go#L133-L137

New query:
https://github.com/gravitational/teleport/blob/ce456d8a7db505c0ab11cbd0c2416f7c4c1719f5/lib/backend/postgres/tx.go#L134-L141

Performance improved from 7 minutes to 76 ms on a 100K record dataset.
```
OLD QUERY
teleport=# explain analyze select key,id from item where (key,id) not in (
        select key,id from lease where (expires is null or expires > now())
        union
        select key,id from event
);
                                                                         QUERY PLAN
------------------------------------------------------------------------------------------------------------------------------------------------------------
 Seq Scan on item  (cost=14489.54..114117475.54 rows=50000 width=19) (actual time=422813.252..422813.254 rows=0 loops=1)
   Filter: (NOT (SubPlan 1))
   Rows Removed by Filter: 100000
   SubPlan 1
     ->  Materialize  (cost=14489.54..16521.56 rows=100001 width=40) (actual time=0.002..1.936 rows=50000 loops=100000)
           ->  Unique  (cost=14489.54..15239.55 rows=100001 width=40) (actual time=90.174..105.293 rows=100000 loops=1)
                 ->  Sort  (cost=14489.54..14739.55 rows=100001 width=40) (actual time=90.173..96.160 rows=100000 loops=1)
                       Sort Key: lease.key, lease.id
                       Sort Method: external merge  Disk: 3336kB
                       ->  Append  (cost=8.60..3447.63 rows=100001 width=40) (actual time=0.038..27.681 rows=100000 loops=1)
                             ->  Bitmap Heap Scan on lease  (cost=8.60..12.62 rows=1 width=19) (actual time=0.027..0.028 rows=0 loops=1)
                                   Recheck Cond: ((expires IS NULL) OR (expires > now()))
                                   ->  BitmapOr  (cost=8.60..8.60 rows=1 width=0) (actual time=0.025..0.025 rows=0 loops=1)
                                         ->  Bitmap Index Scan on lease_expires  (cost=0.00..4.30 rows=1 width=0) (actual time=0.020..0.020 rows=0 loops=1)
                                               Index Cond: (expires IS NULL)
                                         ->  Bitmap Index Scan on lease_expires  (cost=0.00..4.30 rows=1 width=0) (actual time=0.002..0.002 rows=0 loops=1)
                                               Index Cond: (expires > now())
                             ->  Seq Scan on event  (cost=0.00..1935.00 rows=100000 width=19) (actual time=0.009..19.800 rows=100000 loops=1)
 Planning Time: 0.413 ms
 Execution Time: 422814.115 ms

NEW QUERY
teleport=# explain analyze select key,id from item i left join lease l using (key,id) left join event e using (key,id) where e.key is null and (l.key is null or l.expires < now());
                                                            QUERY PLAN
----------------------------------------------------------------------------------------------------------------------------------
 Nested Loop Left Join  (cost=4021.42..8265.51 rows=1 width=19) (actual time=76.165..76.167 rows=0 loops=1)
   Filter: ((l.key IS NULL) OR (l.expires < now()))
   ->  Hash Anti Join  (cost=4021.00..8265.00 rows=1 width=19) (actual time=76.165..76.166 rows=0 loops=1)
         Hash Cond: ((i.key = e.key) AND (i.id = e.id))
         ->  Seq Scan on item i  (cost=0.00..1736.00 rows=100000 width=19) (actual time=0.025..8.708 rows=100000 loops=1)
         ->  Hash  (cost=1935.00..1935.00 rows=100000 width=19) (actual time=38.984..38.984 rows=100000 loops=1)
               Buckets: 65536  Batches: 2  Memory Usage: 3234kB
               ->  Seq Scan on event e  (cost=0.00..1935.00 rows=100000 width=19) (actual time=0.009..14.667 rows=100000 loops=1)
   ->  Index Scan using lease_pk on lease l  (cost=0.42..0.49 rows=1 width=27) (never executed)
         Index Cond: (key = i.key)
         Filter: (i.id = id)
 Planning Time: 0.875 ms
 Execution Time: 76.285 ms
```