@jimbishopp 2022-03-25
@Tener 
> we could benefit from having a comprehensive backend test suite testing backend correctness and performance

We do! The test suite was very helpful when putting this together, especially by validating the correct error is returned from specific backend functions.

https://github.com/gravitational/teleport/blob/3175fb1e08be6d6d3ccf78276b32a236aa907f62/lib/backend/sqlbk/test.go#L70

https://github.com/gravitational/teleport/blob/3175fb1e08be6d6d3ccf78276b32a236aa907f62/lib/backend/test/suite.go#L105-L112