@Tener 2022-03-25
@jimbishopp

> We do! The test suite was very helpful when putting this together, especially by validating the correct error is returned from specific backend functions. I was able to iterate very quickly in the end because of this.

This is good stuff, thanks for pointing that out. We could go bigger still with the tests to get a sense of the relative speed of individual backends under production-like loads... but this is probably just my inner benchmark junkie speaking 😉